refactor: move participant update logic to a service
