+++ app/services/participant_service.ts
@@ -1,7 +1,10 @@
 import Event from "#models/event";
 import Participant from "#models/participant";
 
-import { CreateParticipantDTO } from "../types/participant_types.js";
+import {
+  CreateParticipantDTO,
+  UpdateParticipantDTO,
+} from "../types/participant_types.js";
 
 export class ParticipantService {
   async createParticipant(
@@ -20,7 +23,6 @@ export class ParticipantService {
       participantAttributes !== undefined &&
       participantAttributes.length > 0
     ) {
-      // Transform permissions to match database schema: event_id instead of eventId
       const transformedAttributes = Object.fromEntries(
         participantAttributes.map((participantAttribute) => [
           participantAttribute.attributeId,
@@ -34,5 +36,50 @@ export class ParticipantService {
     await participant.load("attributes");
 
     return participant;
+  }
+
+  async updateParticipant(
+    eventId: number,
+    participantId: number,
+    updateParticipantDTO: UpdateParticipantDTO,
+  ) {
+    const { participantAttributes, ...updates } = updateParticipantDTO;
+
+    const participant = await Participant.query()
+      .where("id", participantId)
+      .andWhere("event_id", eventId)
+      .firstOrFail();
+
+    participant.merge(updates);
+    await participant.save();
+
+    if (
+      participantAttributes !== undefined &&
+      participantAttributes.length > 0
+    ) {
+      const transformedAttributes = Object.fromEntries(
+        participantAttributes.map((participantAttribute) => [
+          participantAttribute.attributeId,
+          { value: participantAttribute.value },
+        ]),
+      );
+
+      await participant
+        .related("attributes")
+        .sync(transformedAttributes, false);
+    }
+
+    const updatedParticipant = await Participant.query()
+      .where("id", participantId)
+      .where("event_id", eventId)
+      .preload("attributes", (attributesQuery) =>
+        attributesQuery
+          .select("id", "name", "slug")
+          .pivotColumns(["value"])
+          .where("show_in_list", true),
+      )
+      .firstOrFail();
+
+    return updatedParticipant;
   }
 }
